cicd/pyproject.toml
[tool.poetry]
name = "cicd"
version = "0.0.1"
description = "cicd pipeline for terraform deployments."
authors = ["Duncan Haywood"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = "^2.45.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
